// frontend/src/openapi/models/ConnectorShapeCandidate.ts
export interface ConnectorShapeCandidate {
    /**
     *
     * @type {string}
     * @memberof ConnectorShapeCandidate
     */
    name: string;
    /**
     *
     * @type {string}
     * @memberof ConnectorShapeCandidate
     */
    type: string;
    /**
     *
     * @type {number}
     * @memberof ConnectorShapeCandidate
     */
    traces: number;
    /**
     *
     * @type {number}
     * @memberof ConnectorShapeCandidate
     */
    occurrences: number;
    /**
     *
     * @type {number}
     * @memberof ConnectorShapeCandidate
     */
    maxPerTrace: number;
    /**
     *
     * @type {number}
     * @memberof ConnectorShapeCandidate
     */
    modelCalls: number;
    /**
     *
     * @type {boolean}
     * @memberof ConnectorShapeCandidate
     */
    isRoot: boolean;
    /**
     *
     * @type {string}
     * @memberof ConnectorShapeCandidate
     */
    parentName: string | null;
    /**
     *
     * @type {number}
     * @memberof ConnectorShapeCandidate
     */
    score: number;
    /**
     *
     * @type {Array<string>}
     * @memberof ConnectorShapeCandidate
     */
    reasons: Array<string>;
}